# budget-buddy-ui/src/app/services/models/contact-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ContactRequest {
  email?: string;
  id?: number;
  message: string;
  sender?: string;
  subject: string;
}