# budget-buddy-ui/src/app/services/models/contact-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ContactResponse {
  email?: string;
  id?: number;
  message?: string;
  sender?: string;
  state?: string;
  subject?: string;
}