# budget-buddy-ui/src/app/services/models/contact-state-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ContactStateRequest {
  code: string;
  description?: string;
  id?: number;
  name: string;
}