# budget-buddy-ui/src/app/services/models/page-response-contact-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ContactResponse } from '../models/contact-response';
export interface PageResponseContactResponse {
  content?: Array<ContactResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}